[Tool Calls] Inline pending tool label helper
